build success

+++ application/src/main/java/com/amazon/proserve/application/customer/command/ChangeCustomerNameCommand.java
@@ -5,13 +5,12 @@ import lombok.Value;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.PositiveOrZero;
-import java.math.BigInteger;
 
 @Value(staticConstructor = "of")
 public class ChangeCustomerNameCommand {
     @NotNull
     @PositiveOrZero
-    BigInteger customerNo;
+    Integer customerNo;
     @NotBlank
     String newCustomerName;
 }
