+++ interface/src/main/java/com/amazon/proserve/customer/api/dto/ChangeCustomerNameRequest.java
@@ -3,8 +3,6 @@ package com.amazon.proserve.customer.api.dto;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
-import java.math.BigInteger;
-
 @Builder
 @Getter
 @FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
@@ -12,6 +10,6 @@ import java.math.BigInteger;
 @ToString
 @EqualsAndHashCode
 public class ChangeCustomerNameRequest {
-    private BigInteger customerNo;
+    private Integer customerNo;
     private String customerNm;
 }
